main/serializers.py
from rest_framework import serializers
from .models import Course

# serializers.py
from .models import Mentor

class MentorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Mentor
        fields = '__all__'


class CourseSerializer(serializers.ModelSerializer):
    image = serializers.ImageField(use_url=True)

    class Meta:
        model = Course
        fields = '__all__'
